W.I.P. This causes linuxviewer to crash.
The problem is related to debug_create_info.
Using this code a temporary is being used, which is
altered to set the call back function; but that alteration
is lost after creating the instance.

Upon creating the device we get a crash for some
yet unknown reason.

I suspect that a call to `setup` is missing on a second
temporary.