Remove Ambifix::operator()(std::string const&)
Now, the right way to append to the prefix of an ambifix
(Ambifix or AmbifixOwner) is by doing: ".extra" + ambifix,
which returns again an Ambifix (AmbifixOwner).

The problem with Ambifix::operator()(infix) was that it
returns a string from which then erroneous an Ambifix could
be created again where the postfix was moved to the prefix.

One can still insert an infix and get a std::string by calling
Ambifix::object_name(infix).