Workaround for magic_enum bug.
It seems that magic_enum always tries to cast 129 to
enum types that it is used with. Unless we use at least
an unsigned char as underlying type, that results in
a warning when using clang.

In the case of shaderc_shader_kind we can't do that
because that is a C enum.